Refactor: Updated PageLayout and theme structure 1. feat(ui): Create reusable PageLayout component for consistent UI
- Add PageLayout component with theme provider integration
- Implement centered layout option for flexible page structures
- Apply consistent background gradient across all pages

2. enhance(theme): Extend darkTheme with custom values and gradients

- Add custom theme values for consistent styling
- Define gradient colors for background elements
- Improve theme structure for better maintainability

3. refactor(pages): Implement PageLayout across application pages

- Update home page to use PageLayout component
- Refactor authentication pages (login, register, verification)
- Apply consistent layout to password management pages
- Reduce code duplication across page components
