+++ expense-tracker-frontend/src/app/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { Box, Stack, ThemeProvider, } from '@mui/material';
-import darkTheme from '../components/common/theme/darkTheme';
+import { Box, Stack } from '@mui/material';
+import PageLayout from '../components/common/layout/PageLayout';
 import SignUpButton from '../components/homeComponents/SignUpButton';
 import SignInButton from '../components/homeComponents/SignInButton';
 import HeroIcon from '../components/homeComponents/HeroIcon';
@@ -10,28 +10,16 @@ import FooterMessage from '../components/homeComponents/FooterMessage';
 
 export default function Home() {
   return (
-    <ThemeProvider theme={darkTheme}>
-      <Box
-        sx={{
-          minHeight: '100vh',
-          background: 'linear-gradient(135deg, #0f172a 0%, #1e293b 100%)',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          p: 2
-        }}
-      >
-        <Box sx={{ textAlign: 'center', maxWidth: 600 }}>
-          <HeroIcon />
-          <HeroText />
-          <Stack direction={{ xs: 'column', sm: 'row' }} spacing={3} justifyContent="center">
-            <SignInButton />
-            <SignUpButton />
-          </Stack>
-          <FooterMessage />
-        </Box>
+    <PageLayout>
+      <Box sx={{ textAlign: 'center', maxWidth: 600 }}>
+        <HeroIcon />
+        <HeroText />
+        <Stack direction={{ xs: 'column', sm: 'row' }} spacing={3} justifyContent="center">
+          <SignInButton />
+          <SignUpButton />
+        </Stack>
+        <FooterMessage />
       </Box>
-    </ThemeProvider>
+    </PageLayout>
   );
 }
-
