+++ expense-tracker-frontend/src/app/register/page.tsx
@@ -1,24 +1,12 @@
 'use client';
 
-import { Box, ThemeProvider } from '@mui/material';
-import darkTheme from '../../components/common/theme/darkTheme';
-import RegisterCard from '../../components/registerComponents/RegisterCard';
+import PageLayout from '@/components/common/layout/PageLayout';
+import RegisterCard from '@/components/registerComponents/RegisterCard';
 
 const UserRegistration = () => (
-  <ThemeProvider theme={darkTheme}>
-    <Box
-      sx={{
-        minHeight: '100vh',
-        background: 'linear-gradient(135deg, #0f172a 0%, #1e293b 100%)',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        p: 2,
-      }}
-    >
-      <RegisterCard />
-    </Box>
-  </ThemeProvider>
+  <PageLayout>
+    <RegisterCard />
+  </PageLayout>
 );
 
 export default UserRegistration;
