+++ expense-tracker-frontend/src/app/reset-password/page.tsx
@@ -1,25 +1,13 @@
 'use client';
 
-import { Box, ThemeProvider } from '@mui/material';
-import darkTheme from '../../components/common/theme/darkTheme';
-import ResetPasswordCard from '../../components/resetPasswordComponents/ResetPasswordCard';
+import PageLayout from '@/components/common/layout/PageLayout';
+import ResetPasswordCard from '@/components/resetPasswordComponents/ResetPasswordCard';
 
 const ResetPasswordPage = () => {
   return (
-    <ThemeProvider theme={darkTheme}>
-      <Box
-        sx={{
-          minHeight: '100vh',
-          background: 'linear-gradient(135deg, #0f172a 0%, #1e293b 100%)',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          p: 2,
-        }}
-      >
-        <ResetPasswordCard />
-      </Box>
-    </ThemeProvider>
+    <PageLayout>
+      <ResetPasswordCard />
+    </PageLayout>
   );
 };
