+++ expense-tracker-frontend/src/components/common/PasswordTooltip.tsx
